Problem 10. This was a bit different!
I had an idea of using a visualizing tool, but ascii art was good
enough. I came up with the solution when I looked at the problem
in the morning but it was in the evening I actually took the time
to write the code for it.

Subproblem 2 was the easiest so far. My program already printed it
so it was solved in a couple of seconds. That is good. Because
I need to sleep.

This is so fun and addictive!

Time: 10511