Day 25 2020
The final puzzle of Advent Of Code this year!

This was about finding out the private part of a private/public key
pair. Very straight forward. I was sort of preparing for part 2, and was
expecting something harder in the end. But this was very nice ending I
think!

Kind problems this year, except for a very few ones. I will try to
continue solving problems from other years to be learn more in general
and hopefully also try other languages.

The primary goal for this year was to solve more problems than the 2
previous years which I have done with great margin.
The secondary goal was to solve all problems which also has been done.
The third goal was to be on the top 10'at the Kodsnack leaderboard but I
ended up at position 11 so I need to be faster next year.

See you next year!