2022 Day 13: Distress Signal
Today it was about parsing lists consisting of numbers and other lists.
When the parsing was done the lists had to be sorted after a few rules.

Today I am home sick so I did not go up in the morning to solve this
problem. Later during the day I was able to gather the needed energy.