Day 8: Resonant Collinearity
This Sunday I decided to sleep instead of walking up instead at 6. I solved it
around lunch time instead.

The tricky part (for me) was to figure out if 3 points were colinear or not. I
spent some time thinking about this before actually googling for a solution. I
usually do not do that but that is how I did it today.